+++ src/main/java/org/cardanofoundation/explorer/api/model/response/pool/projection/PoolListProjection.java
@@ -24,5 +24,5 @@ public interface PoolListProjection {
 
   Integer getVotingPower();
 
-  Integer getGovernanceParticipationRate();
+  Double getGovernanceParticipationRate();
 }
